WebSocket polling added
New mg_poll_websocket() function, so instead of waiting for websocket_data
event, you can enter your own polling loop. A sensible place to do this is
within the websocket_ready event, so when your work is done you can simply
return 0 to close the connection, or 1 to enter to the default read loop.

Modified read_websocket() and handle_websocket_request() to create this
functionality.